refactor: 이메일 인증 모듈을 user.infrastructure.email로 이동 (#58)
