+++ backend/src/main/java/com/pigma/harusari/user/infrastructure/email/service/EmailAuthService.java
@@ -1,4 +1,4 @@
-package com.pigma.harusari.common.email.service;
+package com.pigma.harusari.user.infrastructure.email.service;
 
 public interface EmailAuthService {
